test: split DBF raw memo payload regressions
