+++ CHANGELOG.md
@@ -9433,4 +9433,10 @@ passes `1/1`.
   unaffected.
 - 2026-08-24: Corrected the recovered `EVENTHANDLER()` traceability row so its
   rendered status and issue links remain in their defined table columns; its
-  existing hazard constraints now remain with the verification boundary.
+   existing hazard constraints now remain with the verification boundary.
+
+- 2026-08-23: Continued #2564 test-module refactoring by moving the DBF
+  empty-decoding raw memo-payload and binary picture `PACK MEMO` preservation
+  regressions into `test_dbf_table_raw_memo_payloads.cpp`. The existing
+  `test_dbf_table` executable, invocation order, byte-level assertions, and
+   DBF behavior remain unchanged.
